Time: 3 ms (46.53%), Space: 7.5 MB (99.07%) - LeetHub

+++ 0058-length-of-last-word/0058-length-of-last-word.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int lengthOfLastWord(string s) {
-        int length_s1=0,length_s2;
+        short length_s1=0,length_s2;
         for(int i=0;i<s.length();i++){
             if(s[i]!=32)
             {
